Change loop_rate to support fractional values
Previously, loop_rate was limited to integer values, restricting the Rate
object to whole-second intervals. By changing the data type to double, we
enable fractional  rates, allowing the Rate object to sleep for
intervals longer than 1 second or for precise fractional durations.

+++ include/adi_iio/iio_attr_topic.hpp
@@ -47,7 +47,7 @@ public:
 
   IIOAttrTopic(
     std::shared_ptr<IIONode> nh, std::string topicName, std::string attrPath,
-    topicType_t type, int loopRate);
+    topicType_t type, double loopRate);
   virtual ~IIOAttrTopic();
 
   void publishingLoop();
